requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bilibili-api-python==14.2.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
Pillow==9.4.0
pip==22.3.1
pyasn1==0.4.8
pydantic==1.10.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
wheel==0.37.1
yarl==1.8.2
